Hot_order_agent/requirements.txt

streamlit
pandas
python-dotenv
